ref #264

+++ lib/cache-db/src/main/kotlin/vdi/component/db/cache/sql/select/select-broken-imports.kt
@@ -12,7 +12,7 @@ private fun sqlBody(
   afterFilter:  String,
   orderBy:      String,
   sortOrder:    String,
-  limit:        UByte,
+  limit:        UInt,
   offset:       UInt,
 ) =
 // language=postgresql
